Make sure every gametype has its own folder
