+++ src/game/server/gamecontext.cpp
@@ -31,20 +31,20 @@
 
 #include "entities/character.h"
 #include "gamemodes/DDRace.h"
-#include "gamemodes/instagib/bolofng.h"
-#include "gamemodes/instagib/boomfng.h"
-#include "gamemodes/instagib/fng.h"
+#include "gamemodes/instagib/bolofng/bolofng.h"
+#include "gamemodes/instagib/boomfng/boomfng.h"
+#include "gamemodes/instagib/fng/fng.h"
 #include "gamemodes/instagib/gctf/gctf.h"
-#include "gamemodes/instagib/gdm.h"
-#include "gamemodes/instagib/gtdm.h"
+#include "gamemodes/instagib/gdm/gdm.h"
+#include "gamemodes/instagib/gtdm/gtdm.h"
 #include "gamemodes/instagib/ictf/ictf.h"
-#include "gamemodes/instagib/idm.h"
-#include "gamemodes/instagib/itdm.h"
-#include "gamemodes/instagib/solofng.h"
+#include "gamemodes/instagib/idm/idm.h"
+#include "gamemodes/instagib/itdm/itdm.h"
+#include "gamemodes/instagib/solofng/solofng.h"
 #include "gamemodes/instagib/zcatch/zcatch.h"
 #include "gamemodes/mod.h"
-#include "gamemodes/vanilla/ctf.h"
-#include "gamemodes/vanilla/dm.h"
+#include "gamemodes/vanilla/ctf/ctf.h"
+#include "gamemodes/vanilla/dm/dm.h"
 #include "player.h"
 #include "score.h"
