+++ src/game/server/gamemodes/instagib/fng/fng.cpp
