+++ src/game/server/gamemodes/instagib/gdm/gdm.h
@@ -1,5 +1,5 @@
-#ifndef GAME_SERVER_GAMEMODES_INSTAGIB_GDM_H
-#define GAME_SERVER_GAMEMODES_INSTAGIB_GDM_H
+#ifndef GAME_SERVER_GAMEMODES_INSTAGIB_GDM_GDM_H
+#define GAME_SERVER_GAMEMODES_INSTAGIB_GDM_GDM_H
 
 #include <game/server/gamemodes/instagib/dm.h>
