+++ src/game/server/gamemodes/instagib/idm/idm.h
@@ -1,5 +1,5 @@
-#ifndef GAME_SERVER_GAMEMODES_INSTAGIB_IDM_H
-#define GAME_SERVER_GAMEMODES_INSTAGIB_IDM_H
+#ifndef GAME_SERVER_GAMEMODES_INSTAGIB_IDM_IDM_H
+#define GAME_SERVER_GAMEMODES_INSTAGIB_IDM_IDM_H
 
 #include <game/server/gamemodes/instagib/dm.h>
