+++ src/game/server/gamemodes/instagib/solofng/solofng.cpp
