+++ src/game/server/gamemodes/vanilla/ctf/ctf.h
@@ -1,5 +1,5 @@
-#ifndef GAME_SERVER_GAMEMODES_VANILLA_CTF_H
-#define GAME_SERVER_GAMEMODES_VANILLA_CTF_H
+#ifndef GAME_SERVER_GAMEMODES_VANILLA_CTF_CTF_H
+#define GAME_SERVER_GAMEMODES_VANILLA_CTF_CTF_H
 
 #include <game/server/gamemodes/base_pvp/ctf.h>
